pointing_device_gestures: Adjust cursor glide types
dx & dy can be int16_t, up the accumulated position elements to 32-bits.
Change coef to Q8.

+++ quantum/pointing_device_gestures.h
@@ -26,10 +26,10 @@ typedef struct {
 
 typedef struct {
     uint16_t          trigger_px;
-    float             coef;
+    uint16_t          coef;
     float             v0;
-    int16_t           x;
-    int16_t           y;
+    int32_t           x;
+    int32_t           y;
     uint16_t          z;
     uint16_t          timer;
     uint16_t          interval;
